- Add support for Dynamic Frameworks and Carthage - Convert BVSDKDemo to Swift 3 syntax - Cleanup compiler warnings - Remove deprecated Conversations API from BVSDK 2.x - Update Gimbal.framework to 2.52.1 - Fix use of %lu and %li in some spots - Fix missing negation operator when checking for array type in the BVRatingDistribution - Serveral updates for ci scripts and matrixed travis ci builds

+++ Pod/BVConversations/Submission/Review/BVReviewSubmission.m
@@ -220,7 +220,7 @@
             NSDictionary *json = [NSJSONSerialization JSONObjectWithData:data options:kNilOptions error:&jsonParsingError];
             BVReviewSubmissionErrorResponse* errorResponse = [[BVReviewSubmissionErrorResponse alloc] initWithApiResponse:json]; // fails gracefully
             
-            [[BVLogger sharedLogger] verbose:[NSString stringWithFormat:@"RESPONSE: %@ (%d)", json, statusCode]];
+            [[BVLogger sharedLogger] verbose:[NSString stringWithFormat:@"RESPONSE: %@ (%ld)", json, (long)statusCode]];
             
             if (httpError) {
                 // network error was generated
@@ -282,7 +282,7 @@
                                            @"apiversion": @"5.4",
                                            @"reviewtext": self.reviewText,
                                            @"title": self.reviewTitle,
-                                           @"rating": [NSString stringWithFormat:@"%lu", self.rating],
+                                           @"rating": [NSString stringWithFormat:@"%lu", (unsigned long)self.rating],
                                            @"productId": self.productId
                                        }];
